backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.12.5
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
colorama==0.4.6
